Moves auth files to correct package
