+++ project/app/src/main/java/com/chalmers/tda367/localfeud/authentication/AuthenticatedUser.java
@@ -1,4 +1,4 @@
-package com.chalmers.tda367.localfeud.net.auth;
+package com.chalmers.tda367.localfeud.authentication;
 
 import com.facebook.AccessToken;
